end total Eggs flow

+++ src/main/java/com/wakanda/chickeneggs/eggsRecord/application/api/TotalEggsperChickenperDate.java
@@ -4,5 +4,5 @@ import lombok.Value;
 
 @Value
 public class TotalEggsperChickenperDate {
-    private Double totalEggs;
+    private Integer totalEggs;
 }
